Replace Exception with Custom Error And Fix Lock Release Bug
There was a bug where calling .lock would release the lock even if it never acquired it. The result was that the first time an attempt is made to acquire a busy lock would fail, but then the lock would release. From then on, subsequent calls to #lock would successfully seize the lock when they should not.

Exception is the broadest error class in ruby. By explicitly raising Exception, developers are forced to rescue Exception in order to implement logic for handling timeouts, which is dangerous. Here, it's been replaced with a much safer custom error.